projects/agents/05-agent-memory-general/requirements.txt
agent-framework
mem0
azure-search-documents
python-dotenv
